More method moving
Moved the MIDI player stuff to sound package
