+++ src/main/java/org/runejs/client/ClientScriptRunner.java
@@ -484,11 +484,11 @@ public class ClientScriptRunner extends Node {
                         continue;
                     }
                     if(scriptOpcode == 42) {
-                        scriptIntValues[intValueIndex++] = Class22_Sub1.anIntArray1847[intOperands[scriptIndex]];
+                        scriptIntValues[intValueIndex++] = MovedStatics.anIntArray1847[intOperands[scriptIndex]];
                         continue;
                     }
                     if(scriptOpcode == 43) {
-                        Class22_Sub1.anIntArray1847[intOperands[scriptIndex]] = scriptIntValues[--intValueIndex];
+                        MovedStatics.anIntArray1847[intOperands[scriptIndex]] = scriptIntValues[--intValueIndex];
                         continue;
                     }
                 }
@@ -519,10 +519,10 @@ public class ClientScriptRunner extends Node {
                         if(bool) {
                             MovedStatics.aGameInterface_2116 = childInterface;
                         } else {
-                            Class22_Sub2.aGameInterface_1887 = childInterface;
+                            MovedStatics.aGameInterface_1887 = childInterface;
                         }
                     } else if(scriptOpcode == 101) {
-                        GameInterface gameInterface = !bool ? Class22_Sub2.aGameInterface_1887 : MovedStatics.aGameInterface_2116;
+                        GameInterface gameInterface = !bool ? MovedStatics.aGameInterface_1887 : MovedStatics.aGameInterface_2116;
                         GameInterface childInterface = GameInterface.getInterface(gameInterface.parentId);
                         childInterface.children[BitUtils.bitWiseAND(gameInterface.id, 32767)] = null;
                     } else {
@@ -538,7 +538,7 @@ public class ClientScriptRunner extends Node {
                         gameInterface = GameInterface.getInterface(scriptIntValues[--intValueIndex]);
                         scriptOpcode -= 1000;
                     } else {
-                        gameInterface = bool ? MovedStatics.aGameInterface_2116 : Class22_Sub2.aGameInterface_1887;
+                        gameInterface = bool ? MovedStatics.aGameInterface_2116 : MovedStatics.aGameInterface_1887;
                     }
                     if(scriptOpcode == 1000) {
                         intValueIndex -= 2;
@@ -559,7 +559,7 @@ public class ClientScriptRunner extends Node {
                 } else if(scriptOpcode >= 1100 && scriptOpcode < 1200 || scriptOpcode >= 2100 && scriptOpcode < 2200) {
                     GameInterface gameInterface;
                     if(scriptOpcode < 2000) {
-                        gameInterface = !bool ? Class22_Sub2.aGameInterface_1887 : MovedStatics.aGameInterface_2116;
+                        gameInterface = !bool ? MovedStatics.aGameInterface_1887 : MovedStatics.aGameInterface_2116;
                     } else {
                         gameInterface = GameInterface.getInterface(scriptIntValues[--intValueIndex]);
                         scriptOpcode -= 1000;
@@ -631,7 +631,7 @@ public class ClientScriptRunner extends Node {
                                     if(bool) {
                                         MovedStatics.aGameInterface_2116 = gameInterface.children[i_28_];
                                     } else {
-                                        Class22_Sub2.aGameInterface_1887 = gameInterface.children[i_28_];
+                                        MovedStatics.aGameInterface_1887 = gameInterface.children[i_28_];
                                     }
                                 }
                             } else if(scriptOpcode == 1401) {
@@ -647,7 +647,7 @@ public class ClientScriptRunner extends Node {
                                     if(bool) {
                                         MovedStatics.aGameInterface_2116 = gameInterface;
                                     } else {
-                                        Class22_Sub2.aGameInterface_1887 = gameInterface;
+                                        MovedStatics.aGameInterface_1887 = gameInterface;
                                     }
                                 }
                             } else {
@@ -666,12 +666,12 @@ public class ClientScriptRunner extends Node {
                                     if(bool) {
                                         MovedStatics.aGameInterface_2116 = gameInterface_35_;
                                     } else {
-                                        Class22_Sub2.aGameInterface_1887 = gameInterface_35_;
+                                        MovedStatics.aGameInterface_1887 = gameInterface_35_;
                                     }
                                 }
                             }
                         } else if(scriptOpcode < 1600) {
-                            GameInterface gameInterface = bool ? MovedStatics.aGameInterface_2116 : Class22_Sub2.aGameInterface_1887;
+                            GameInterface gameInterface = bool ? MovedStatics.aGameInterface_2116 : MovedStatics.aGameInterface_1887;
                             if(scriptOpcode == 1500) {
                                 scriptIntValues[intValueIndex++] = gameInterface.currentX;
                             } else if(scriptOpcode == 1501) {
@@ -744,7 +744,7 @@ public class ClientScriptRunner extends Node {
                                             int i_44_ = scriptIntValues[intValueIndex + 1];
                                             Class33.method406(0, i_44_, i_43_, -128);
                                         } else if(scriptOpcode == 3002) {
-                                            GameInterface gameInterface = !bool ? Class22_Sub2.aGameInterface_1887 : MovedStatics.aGameInterface_2116;
+                                            GameInterface gameInterface = !bool ? MovedStatics.aGameInterface_1887 : MovedStatics.aGameInterface_2116;
                                             if(MovedStatics.lastContinueTextWidgetId == -1) {
                                                 PacketBuffer.method517(gameInterface.id & 0x7fff, gameInterface.parentId);
                                                 MovedStatics.lastContinueTextWidgetId = gameInterface.id;
@@ -753,7 +753,7 @@ public class ClientScriptRunner extends Node {
                                             if(scriptOpcode != 3003) {
                                                 break;
                                             }
-                                            GameInterface gameInterface = bool ? MovedStatics.aGameInterface_2116 : Class22_Sub2.aGameInterface_1887;
+                                            GameInterface gameInterface = bool ? MovedStatics.aGameInterface_2116 : MovedStatics.aGameInterface_1887;
                                             int i_45_ = scriptIntValues[--intValueIndex];
                                             Class33.method406(0x7fff & gameInterface.id, i_45_, gameInterface.parentId, -121);
                                         }
@@ -939,7 +939,7 @@ public class ClientScriptRunner extends Node {
                                 }
                             }
                         } else {
-                            GameInterface gameInterface = bool ? MovedStatics.aGameInterface_2116 : Class22_Sub2.aGameInterface_1887;
+                            GameInterface gameInterface = bool ? MovedStatics.aGameInterface_2116 : MovedStatics.aGameInterface_1887;
                             if(scriptOpcode == 1600) {
                                 scriptIntValues[intValueIndex++] = gameInterface.scrollWidth;
                             } else {
@@ -955,7 +955,7 @@ public class ClientScriptRunner extends Node {
                             gameInterface = GameInterface.getInterface(scriptIntValues[--intValueIndex]);
                             scriptOpcode -= 1000;
                         } else {
-                            gameInterface = !bool ? Class22_Sub2.aGameInterface_1887 : MovedStatics.aGameInterface_2116;
+                            gameInterface = !bool ? MovedStatics.aGameInterface_1887 : MovedStatics.aGameInterface_2116;
                         }
                         if(scriptOpcode >= 1300 && scriptOpcode <= 1309 || scriptOpcode >= 1314 && scriptOpcode <= 1317) {
                             String class1 = scriptStringValues[--stringValueIndex];
@@ -1034,7 +1034,7 @@ public class ClientScriptRunner extends Node {
                 } else {
                     GameInterface gameInterface;
                     if(scriptOpcode < 2000) {
-                        gameInterface = !bool ? Class22_Sub2.aGameInterface_1887 : MovedStatics.aGameInterface_2116;
+                        gameInterface = !bool ? MovedStatics.aGameInterface_1887 : MovedStatics.aGameInterface_2116;
                     } else {
                         gameInterface = GameInterface.getInterface(scriptIntValues[--intValueIndex]);
                         scriptOpcode -= 1000;
@@ -1066,7 +1066,7 @@ public class ClientScriptRunner extends Node {
                         if(scriptOpcode != 1203) {
                             break;
                         }
-                        GameInterface desiredInterface = !bool ? MovedStatics.aGameInterface_2116 : Class22_Sub2.aGameInterface_1887;
+                        GameInterface desiredInterface = !bool ? MovedStatics.aGameInterface_2116 : MovedStatics.aGameInterface_1887;
                         gameInterface.anInt2738 = desiredInterface.id;
                     }
                 }
