+++ src/main/java/org/runejs/client/MovedStatics.java
@@ -35,6 +35,7 @@ import org.runejs.client.scene.InteractiveObject;
 import org.runejs.client.scene.Scene;
 import org.runejs.client.scene.SceneCluster;
 import org.runejs.client.scene.util.CollisionMap;
+import org.runejs.client.sound.Class22;
 import org.runejs.client.sound.StaticAudio;
 import org.runejs.client.util.BitUtils;
 import org.runejs.client.util.Signlink;
@@ -172,7 +173,7 @@ public class MovedStatics {
             Class39.aProducingGraphicsBuffer_907 = null;
             GameObjectDefinition.aProducingGraphicsBuffer_2524 = null;
             Class39.aClass40_Sub5_Sub14_Sub4_918 = null;
-            Class22.aClass40_Sub5_Sub14_Sub2Array535 = null;
+            MovedStatics.aClass40_Sub5_Sub14_Sub2Array535 = null;
             Class40_Sub5_Sub17_Sub6.anIntArray3248 = null;
             MovedStatics.loginBoxGraphics = null;
             SceneTile.aClass40_Sub5_Sub14_Sub4_2043 = null;
@@ -188,7 +189,6 @@ public class MovedStatics {
 
     public static int[] menuActionTypes = new int[500];
     public static ProducingGraphicsBuffer loginBoxGraphics;
-    public static int[] anIntArray889 = new int[128];
     public static int anInt892;
     public static boolean showSidePanelRedrawnText = false;
     public static IndexedImage aClass40_Sub5_Sub14_Sub2_1919;
@@ -637,7 +637,7 @@ public class MovedStatics {
         if (RSRuntimeException.anInt1641 > Landscape.anIntArray1168.length) {
             RSRuntimeException.anInt1641 -= Landscape.anIntArray1168.length;
             int i_9_ = (int) (12.0 * Math.random());
-            FramemapDefinition.method879(Class22.aClass40_Sub5_Sub14_Sub2Array535[i_9_]);
+            FramemapDefinition.method879(MovedStatics.aClass40_Sub5_Sub14_Sub2Array535[i_9_]);
         }
         for (int i_10_ = 1; i_10_ < -1 + i; i_10_++) {
             for (int i_11_ = 1; i_11_ < 127; i_11_++) {
@@ -1817,4 +1817,334 @@ public class MovedStatics {
 	public static int anInt2452 = 0;
 	public static int loadingPercent = 0;
 	public static int cameraY;
+
+	public static synchronized void method407(boolean arg0) {
+	    MovedStatics.method1019();
+	}
+
+	public static void method1019() {
+	    if(aClass22_189 != null) {
+	        MovedStatics.method308();   
+            if(RSCanvas.anInt54 > 0) {
+                aClass22_189.method301(256, 0);
+                RSCanvas.anInt54 = 0;
+            }
+            aClass22_189.method306((byte) 101);
+            aClass22_189 = null;
+	    }
+	}
+
+	public static void method309(int varPlayerIndex) {
+	    do {
+	        AnimationSequence.anInt2480 = pulseCycle;
+	        StaticAudio.method1030();
+	        int varPlayerType = VarPlayerDefinition.getDefinition(varPlayerIndex).type;
+	        if(varPlayerType != 0) {
+	            int varPlayerValue = VarPlayerDefinition.varPlayers[varPlayerIndex];
+	            if(varPlayerType == 1) {
+	                if(varPlayerValue == 1) {
+	                    Rasterizer3D.method711(0.9);
+	                    ((Class35) Rasterizer3D.anInterface3_2939).method424(0.9);
+	                }
+	                if(varPlayerValue == 2) {
+	                    Rasterizer3D.method711(0.8);
+	                    ((Class35) Rasterizer3D.anInterface3_2939).method424(0.8);
+	                }
+	                if(varPlayerValue == 3) {
+	                    Rasterizer3D.method711(0.7);
+	                    ((Class35) Rasterizer3D.anInterface3_2939).method424(0.7);
+	                }
+	                if(varPlayerValue == 4) {
+	                    Rasterizer3D.method711(0.6);
+	                    ((Class35) Rasterizer3D.anInterface3_2939).method424(0.6);
+	                }
+	                GameObject.clearImageCache();
+	                clearScreen = true;
+	            }
+	            if(varPlayerType == 3) {
+	                int i_22_ = 0;
+	                if(varPlayerValue == 0)
+	                    i_22_ = 255;
+	                if(varPlayerValue == 1)
+	                    i_22_ = 192;
+	                if(varPlayerValue == 2)
+	                    i_22_ = 128;
+	                if(varPlayerValue == 3)
+	                    i_22_ = 64;
+	                if(varPlayerValue == 4)
+	                    i_22_ = 0;
+	                if(i_22_ != RSCanvas.musicVolume) {
+	                    if(RSCanvas.musicVolume != 0 || StaticAudio.currentSongId == -1) {
+	                        if(i_22_ == 0) {
+	                            StaticAudio.method402(false);
+	                            StaticAudio.songTimeout = 0;
+	                        } else
+	                            method456(i_22_);
+	                    } else {
+	                        StaticAudio.method414(false, 0, StaticAudio.currentSongId, i_22_, 0, CacheArchive.musicCacheArchive);
+	                        StaticAudio.songTimeout = 0;
+	                    }
+	                    RSCanvas.musicVolume = i_22_;
+	                }
+	            }
+	            if(varPlayerType == 9)
+	                Class43.bankInsertMode = varPlayerValue;
+	            if(varPlayerType == 10) {
+	                if(varPlayerValue == 0)
+	                    StaticAudio.soundEffectVolume = 127;
+	                if(varPlayerValue == 1)
+	                    StaticAudio.soundEffectVolume = 96;
+	                if(varPlayerValue == 2)
+	                    StaticAudio.soundEffectVolume = 64;
+	                if(varPlayerValue == 3)
+	                    StaticAudio.soundEffectVolume = 32;
+	                if(varPlayerValue == 4)
+	                    StaticAudio.soundEffectVolume = 0;
+	            }
+	            if(varPlayerType == 8) {
+	                ChatBox.redrawChatbox = true;
+	                CollisionMap.anInt165 = varPlayerValue;
+	            }
+	            if(varPlayerType == 4) {
+	                if(varPlayerValue == 0)
+	                    anInt200 = 127;
+	                if(varPlayerValue == 1)
+	                    anInt200 = 96;
+	                if(varPlayerValue == 2)
+	                    anInt200 = 64;
+	                if(varPlayerValue == 3)
+	                    anInt200 = 32;
+	                if(varPlayerValue == 4)
+	                    anInt200 = 0;
+	            }
+	            if(varPlayerType == 6)
+	                anInt2280 = varPlayerValue;
+	            if(varPlayerType != 5)
+	                break;
+	            ProducingGraphicsBuffer.oneMouseButton = varPlayerValue;
+	        }
+	
+	        break;
+	    } while(false);
+	}
+
+	public static String method307(Buffer arg0, int arg1, int arg2) {
+	    try {
+	        if(arg1 != -1)
+	            MovedStatics.aBooleanArray548 = null;
+	        int length = arg0.getSmart();
+	        if(length > arg2)
+	            length = arg2;
+	        byte[] chars = new byte[length];
+	        arg0.currentPosition += IdentityKit.aHuffmanEncoding_2590.method1023(arg0.buffer, length, 0, chars, arg0.currentPosition, -1);
+	        return new String(chars);
+	    } catch(Exception exception) {
+	        return English.cabbage;
+	    }
+	}
+
+	public static void method308() {
+	    RSString.method56(false, null, 0);
+	}
+
+	public static void method305() {
+	//        if(ScreenController.frameMode == ScreenMode.FIXED){
+	
+	            RSCanvas.chatBoxImageProducer.prepareRasterizer();
+	//        }
+	        Class44.chatboxBackgroundImage.drawImage(0, 0);
+	        chatboxLineOffsets = Rasterizer3D.setLineOffsets(chatboxLineOffsets);
+	    }
+
+	public static void method299(byte arg0, int arg1) {
+	    int[] is = minimapImage.pixels;
+	    int i = is.length;
+	    for(int i_0_ = 0; i > i_0_; i_0_++)
+	        is[i_0_] = 0;
+	    for(int i_1_ = 1; i_1_ < 103; i_1_++) {
+	        int i_2_ = 24628 + (-(512 * i_1_) + 52736) * 4;
+	        for(int i_3_ = 1; i_3_ < 103; i_3_++) {
+	            if((0x18 & OverlayDefinition.tile_flags[arg1][i_3_][i_1_]) == 0)
+	                Npc.currentScene.method96(is, i_2_, 512, arg1, i_3_, i_1_);
+	            if(arg1 < 3 && (OverlayDefinition.tile_flags[1 + arg1][i_3_][i_1_] & 0x8) != 0)
+	                Npc.currentScene.method96(is, i_2_, 512, 1 + arg1, i_3_, i_1_);
+	            i_2_ += 4;
+	        }
+	    }
+	    minimapImage.method723();
+	    int i_4_ = (-10 + (int) (Math.random() * 20.0) + 238 << 8) + (228 + (int) (Math.random() * 20.0) << 16) + 238 + (int) (20.0 * Math.random()) + -10;
+	    int i_5_ = -10 + (int) (20.0 * Math.random()) + 238 << 16;
+	    for(int i_6_ = 1; i_6_ < 103; i_6_++) {
+	        for(int i_7_ = 1; i_7_ < 103; i_7_++) {
+	            if((OverlayDefinition.tile_flags[arg1][i_7_][i_6_] & 0x18) == 0)
+	                Actor.method781(1850, arg1, i_7_, i_4_, i_5_, i_6_);
+	            if(arg1 < 3 && (0x8 & OverlayDefinition.tile_flags[1 + arg1][i_7_][i_6_]) != 0)
+	                Actor.method781(1850, 1 + arg1, i_7_, i_4_, i_5_, i_6_);
+	        }
+	    }
+	    GameObject.minimapHintCount = 0;
+	    if(arg0 < 24)
+	        Player.trackedPlayerAppearanceCache = null;
+	    for(int i_8_ = 0; i_8_ < 104; i_8_++) {
+	        for(int i_9_ = 0; i_9_ < 104; i_9_++) {
+	            int i_10_ = Npc.currentScene.getFloorDecorationHash(Player.worldLevel, i_8_, i_9_);
+	            if(i_10_ != 0) {
+	                i_10_ = 0x7fff & i_10_ >> 14;
+	                int i_11_ = GameObjectDefinition.getDefinition(i_10_).icon;
+	                if(i_11_ >= 0) {
+	                    int i_12_ = i_9_;
+	                    int i_13_ = i_8_;
+	                    if(i_11_ != 22 && i_11_ != 29 && i_11_ != 34 && i_11_ != 36 && i_11_ != 46 && i_11_ != 47 && i_11_ != 48) {
+	                        int[][] is_14_ = Landscape.currentCollisionMap[Player.worldLevel].clippingData;
+	                        for(int i_15_ = 0; i_15_ < 10; i_15_++) {
+	                            int i_16_ = (int) (Math.random() * 4.0);
+	                            if(i_16_ == 0 && i_13_ > 0 && i_13_ > -3 + i_8_ && (is_14_[-1 + i_13_][i_12_] & 0x1280108) == 0)
+	                                i_13_--;
+	                            if(i_16_ == 1 && i_13_ < 103 && i_13_ < i_8_ + 3 && (is_14_[i_13_ + 1][i_12_] & 0x1280180) == 0)
+	                                i_13_++;
+	                            if(i_16_ == 2 && i_12_ > 0 && i_12_ > -3 + i_9_ && (is_14_[i_13_][i_12_ - 1] & 0x1280102) == 0)
+	                                i_12_--;
+	                            if(i_16_ == 3 && i_12_ < 103 && 3 + i_9_ > i_12_ && (0x1280120 & is_14_[i_13_][1 + i_12_]) == 0)
+	                                i_12_++;
+	                        }
+	                    }
+	                    MouseHandler.minimapHint[GameObject.minimapHintCount] = aClass40_Sub5_Sub14_Sub4Array296[i_11_];
+	                    Actor.minimapHintX[GameObject.minimapHintCount] = i_13_;
+	                    LinkedList.minimapHintY[GameObject.minimapHintCount] = i_12_;
+	                    GameObject.minimapHintCount++;
+	                }
+	            }
+	        }
+	    }
+	}
+
+	public static boolean[] aBooleanArray548 = new boolean[]{true, true, true, true, true, true, true, true, true, true, true, true, true, true, true, true, true, true, true, true, true, false, true, false, false};
+	public static IndexedImage[] aClass40_Sub5_Sub14_Sub2Array535;
+	public static int anInt537 = 0;
+	public static boolean membersWorld = false;
+	public static boolean accountFlagged = false;
+	public static int anInt545;
+	public static GameInterface aGameInterface_1887;
+	public static int duplicateClickCount = 0;
+
+	public static IndexedImage[] method315() {
+	    IndexedImage[] class40_sub5_sub14_sub2s = new IndexedImage[anInt2581];
+	    for(int i = 0; anInt2581 > i; i++) {
+	        IndexedImage class40_sub5_sub14_sub2 = class40_sub5_sub14_sub2s[i] = new IndexedImage();
+	        class40_sub5_sub14_sub2.maxWidth = ItemDefinition.anInt2846;
+	        class40_sub5_sub14_sub2.maxHeight = GameShell.anInt31;
+	        class40_sub5_sub14_sub2.xDrawOffset = Class57.anIntArray1347[i];
+	        class40_sub5_sub14_sub2.yDrawOffset = Actor.anIntArray3111[i];
+	        class40_sub5_sub14_sub2.imgWidth = Class17.anIntArray456[i];
+	        class40_sub5_sub14_sub2.imgHeight = Npc.anIntArray3312[i];
+	        class40_sub5_sub14_sub2.palette = Buffer.anIntArray1972;
+	        class40_sub5_sub14_sub2.imgPixels = GroundItemTile.aByteArrayArray1370[i];
+	    }
+	    ActorDefinition.method569();
+	    return class40_sub5_sub14_sub2s;
+	}
+
+	public static ImageRGB[] method319(byte arg0) {
+	    ImageRGB[] class40_sub5_sub14_sub4s = new ImageRGB[anInt2581];
+	    if(arg0 != -62)
+	        return null;
+	    for(int i = 0; i < anInt2581; i++) {
+	        ImageRGB class40_sub5_sub14_sub4 = class40_sub5_sub14_sub4s[i] = new ImageRGB();
+	        class40_sub5_sub14_sub4.maxWidth = ItemDefinition.anInt2846;
+	        class40_sub5_sub14_sub4.maxHeight = GameShell.anInt31;
+	        class40_sub5_sub14_sub4.offsetX = Class57.anIntArray1347[i];
+	        class40_sub5_sub14_sub4.offsetY = Actor.anIntArray3111[i];
+	        class40_sub5_sub14_sub4.imageWidth = Class17.anIntArray456[i];
+	        class40_sub5_sub14_sub4.imageHeight = Npc.anIntArray3312[i];
+	        byte[] is = GroundItemTile.aByteArrayArray1370[i];
+	        int i_4_ = class40_sub5_sub14_sub4.imageHeight * class40_sub5_sub14_sub4.imageWidth;
+	        class40_sub5_sub14_sub4.pixels = new int[i_4_];
+	        for(int i_5_ = 0; i_5_ < i_4_; i_5_++)
+	            class40_sub5_sub14_sub4.pixels[i_5_] = Buffer.anIntArray1972[BitUtils.bitWiseAND(255, is[i_5_])];
+	    }
+	    ActorDefinition.method569();
+	    return class40_sub5_sub14_sub4s;
+	}
+
+	public static void method311(Component arg1) {
+	    arg1.removeKeyListener(Class59.keyFocusListener);
+	    arg1.removeFocusListener(Class59.keyFocusListener);
+	}
+
+	public static void method450(byte arg0) {
+	    if (Player.headIconDrawType == 2) {
+	        if (arg0 >= -28)
+	            method445(-128);
+	        MovedStatics.method312(2 * ActorDefinition.anInt2404, Class35.anInt1730 + (-Class26.baseY + anInt175 << 7), (ProducingGraphicsBuffer.anInt1637 + -baseX << 7) + Landscape.anInt1170, 4976905);
+	        if (ISAAC.anInt522 > -1 && pulseCycle % 20 < 10)
+	            aClass40_Sub5_Sub14_Sub4Array2567[0].drawImage(ISAAC.anInt522 + -12, -28 + Class44.anInt1048);
+	    }
+	}
+
+	public static void method312(int arg0, int arg1, int arg2, int arg3) {
+	    if(arg2 < 128 || arg1 < 128 || arg2 > 13056 || arg1 > 13056) {
+	        Class44.anInt1048 = -1;
+	        ISAAC.anInt522 = -1;
+	    } else {
+	        int i = Class37.getFloorDrawHeight(Player.worldLevel, arg2, arg1) + -arg0;
+	        arg1 -= cameraY;
+	        i -= SceneCluster.cameraZ;
+	        int i_1_ = Model.COSINE[Class26.anInt627];
+	        int i_2_ = Model.SINE[Class26.anInt627];
+	        arg2 -= Class12.cameraX;
+	        int i_3_ = Model.SINE[ProducingGraphicsBuffer_Sub1.anInt2210];
+	        int i_4_ = Model.COSINE[ProducingGraphicsBuffer_Sub1.anInt2210];
+	        int i_5_ = arg1 * i_3_ + arg2 * i_4_ >> 16;
+	        arg1 = i_4_ * arg1 - arg2 * i_3_ >> 16;
+	        if(arg3 != 4976905)
+	            English.password = null;
+	        arg2 = i_5_;
+	        i_5_ = i * i_1_ - arg1 * i_2_ >> 16;
+	        arg1 = arg1 * i_1_ + i * i_2_ >> 16;
+	        i = i_5_;
+	        if(arg1 < 50) {
+	            Class44.anInt1048 = -1;
+	            ISAAC.anInt522 = -1;
+	        } else {
+	            if(ScreenController.frameMode == ScreenMode.FIXED){
+	                ISAAC.anInt522 = 256 + (arg2 << 9) / arg1;
+	                Class44.anInt1048 = (i << 9) / arg1 + 167;
+	            } else {
+	                ISAAC.anInt522 = ScreenController.drawWidth/2 + (arg2 << 9) / arg1;
+	                Class44.anInt1048 = (i << 9) / arg1 +  ScreenController.drawHeight/2;
+	            }
+	
+	        }
+	    }
+	}
+
+	public static void method313() {
+	    for(int i = -1; Player.localPlayerCount > i; i++) {
+	        int i_6_;
+	        if(i == -1)
+	            i_6_ = 2047;
+	        else
+	            i_6_ = Player.trackedPlayerIndices[i];
+	        Player class40_sub5_sub17_sub4_sub1 = Player.trackedPlayers[i_6_];
+	        if(class40_sub5_sub17_sub4_sub1 != null && class40_sub5_sub17_sub4_sub1.anInt3078 > 0) {
+	            class40_sub5_sub17_sub4_sub1.anInt3078--;
+	            if(class40_sub5_sub17_sub4_sub1.anInt3078 == 0)
+	                class40_sub5_sub17_sub4_sub1.forcedChatMessage = null;
+	        }
+	    }
+	    for(int i_7_ = 0; i_7_ < Player.npcCount; i_7_++) {
+	        int i_8_ = Player.npcIds[i_7_];
+	        Npc class40_sub5_sub17_sub4_sub2 = Player.npcs[i_8_];
+	        if(class40_sub5_sub17_sub4_sub2 != null && class40_sub5_sub17_sub4_sub2.anInt3078 > 0) {
+	            class40_sub5_sub17_sub4_sub2.anInt3078--;
+	            if(class40_sub5_sub17_sub4_sub2.anInt3078 == 0)
+	                class40_sub5_sub17_sub4_sub2.forcedChatMessage = null;
+	        }
+	    }
+	}
+
+	public static int anInt1856;
+	public static int[] anIntArray1847 = new int[2000];
+	public static int[] anIntArray1846 = new int[5];
+	public static long aLong1841;
 }
