+++ src/main/java/org/runejs/client/frame/ChatBox.java
@@ -40,7 +40,7 @@ public class ChatBox {
             '$', '%', '"', '[', ']', '_', '{', '}', '/', '|'};
     public static void renderChatbox() {
         MovedStatics.showChatPanelRedrawnText = true;
-        Class22.method305();
+        MovedStatics.method305();
         if(messagePromptRaised) {
             TypeFace.fontBold.drawStringLeft(Native.aClass1_1563, 239, 40, 0);
             TypeFace.fontBold.drawStringLeft(chatMessage + "*", 239, 60, 128);
