+++ src/main/java/org/runejs/client/scene/tile/FloorDecoration.java
@@ -39,7 +39,7 @@ public class FloorDecoration {
     public int z;
 
     public static void method342(int arg1, Actor arg2) {
-        Class22_Sub1.method312(arg1, arg2.worldY, arg2.worldX, 4976905);
+        MovedStatics.method312(arg1, arg2.worldY, arg2.worldX, 4976905);
     }
 
     public static void constructMapRegion(boolean arg0, int arg1) {
