+++ src/main/java/org/runejs/client/sound/Class22_Sub2.java
@@ -1,67 +1,19 @@
-package org.runejs.client;
+package org.runejs.client.sound;
 
-import org.runejs.client.cache.def.ActorDefinition;
-import org.runejs.client.cache.def.ItemDefinition;
-import org.runejs.client.cache.media.ImageRGB;
-import org.runejs.client.cache.media.IndexedImage;
-import org.runejs.client.cache.media.gameInterface.GameInterface;
-import org.runejs.client.io.Buffer;
+import org.runejs.client.Class44;
 import org.runejs.client.language.Native;
-import org.runejs.client.media.renderable.actor.Actor;
-import org.runejs.client.media.renderable.actor.Npc;
-import org.runejs.client.scene.GroundItemTile;
-import org.runejs.client.util.BitUtils;
 
 public abstract class Class22_Sub2 extends Class22 {
-    public static GameInterface aGameInterface_1887;
-    public static int duplicateClickCount = 0;
     public static int[] anIntArray774 = new int[]{12800, 12800, 12800, 12800, 12800, 12800, 12800, 12800, 12800, 12800, 12800, 12800, 12800, 12800, 12800, 12800};
 
-
-    public static IndexedImage[] method315() {
-        IndexedImage[] class40_sub5_sub14_sub2s = new IndexedImage[MovedStatics.anInt2581];
-        for(int i = 0; MovedStatics.anInt2581 > i; i++) {
-            IndexedImage class40_sub5_sub14_sub2 = class40_sub5_sub14_sub2s[i] = new IndexedImage();
-            class40_sub5_sub14_sub2.maxWidth = ItemDefinition.anInt2846;
-            class40_sub5_sub14_sub2.maxHeight = GameShell.anInt31;
-            class40_sub5_sub14_sub2.xDrawOffset = Class57.anIntArray1347[i];
-            class40_sub5_sub14_sub2.yDrawOffset = Actor.anIntArray3111[i];
-            class40_sub5_sub14_sub2.imgWidth = Class17.anIntArray456[i];
-            class40_sub5_sub14_sub2.imgHeight = Npc.anIntArray3312[i];
-            class40_sub5_sub14_sub2.palette = Buffer.anIntArray1972;
-            class40_sub5_sub14_sub2.imgPixels = GroundItemTile.aByteArrayArray1370[i];
-        }
-        ActorDefinition.method569();
-        return class40_sub5_sub14_sub2s;
-    }
-
-    public static ImageRGB[] method319(byte arg0) {
-        ImageRGB[] class40_sub5_sub14_sub4s = new ImageRGB[MovedStatics.anInt2581];
-        if(arg0 != -62)
-            return null;
-        for(int i = 0; i < MovedStatics.anInt2581; i++) {
-            ImageRGB class40_sub5_sub14_sub4 = class40_sub5_sub14_sub4s[i] = new ImageRGB();
-            class40_sub5_sub14_sub4.maxWidth = ItemDefinition.anInt2846;
-            class40_sub5_sub14_sub4.maxHeight = GameShell.anInt31;
-            class40_sub5_sub14_sub4.offsetX = Class57.anIntArray1347[i];
-            class40_sub5_sub14_sub4.offsetY = Actor.anIntArray3111[i];
-            class40_sub5_sub14_sub4.imageWidth = Class17.anIntArray456[i];
-            class40_sub5_sub14_sub4.imageHeight = Npc.anIntArray3312[i];
-            byte[] is = GroundItemTile.aByteArrayArray1370[i];
-            int i_4_ = class40_sub5_sub14_sub4.imageHeight * class40_sub5_sub14_sub4.imageWidth;
-            class40_sub5_sub14_sub4.pixels = new int[i_4_];
-            for(int i_5_ = 0; i_5_ < i_4_; i_5_++)
-                class40_sub5_sub14_sub4.pixels[i_5_] = Buffer.anIntArray1972[BitUtils.bitWiseAND(255, is[i_5_])];
-        }
-        ActorDefinition.method569();
-        return class40_sub5_sub14_sub4s;
-    }
 
     public static int method323(int arg0) {
         int i = anIntArray774[arg0];
         i = (i * Class44.anInt1040 >> 8) * i;
         return (int) (Math.sqrt(i) + 0.5);
     }
+
+	public static int[] anIntArray889 = new int[128];
 
     public abstract void method314(int i, int i_0_, int i_1_, long l);
 
@@ -98,8 +50,8 @@ public abstract class Class22_Sub2 extends Class22 {
 
     public void method321(byte arg0, long arg1) {
         for(int i = 0; i < 128; i++) {
-            int i_6_ = MovedStatics.anIntArray889[i];
-            MovedStatics.anIntArray889[i] = 0;
+            int i_6_ = Class22_Sub2.anIntArray889[i];
+            Class22_Sub2.anIntArray889[i] = 0;
             for(int i_7_ = 0; i_7_ < 16; i_7_++) {
                 if((i_6_ & 1 << i_7_) != 0)
                     method314(144 + i_7_, i, 0, arg1);
@@ -124,14 +76,14 @@ public abstract class Class22_Sub2 extends Class22 {
     public boolean method322(int arg0, int arg1, int arg2, long arg3) {
         if((arg0 & 0xe0) == 128) {
             int i = 1 << (arg0 & 0xf);
-            int i_8_ = MovedStatics.anIntArray889[arg1];
+            int i_8_ = Class22_Sub2.anIntArray889[arg1];
             if(arg0 >= 144 && arg2 != 0) {
                 if((i_8_ & i) != 0)
                     method314(arg0, arg1, 0, arg3);
                 else
-                    MovedStatics.anIntArray889[arg1] = i_8_ | i;
+                    Class22_Sub2.anIntArray889[arg1] = i_8_ | i;
             } else
-                MovedStatics.anIntArray889[arg1] = i_8_ & (i ^ 0xffffffff);
+                Class22_Sub2.anIntArray889[arg1] = i_8_ & (i ^ 0xffffffff);
             return false;
         }
         if((arg0 & 0xf0) == 176) {
