+++ src/main/java/org/runejs/client/util/Signlink.java
@@ -1,8 +1,8 @@
 package org.runejs.client.util;
 
 import org.runejs.client.SizedAccessFile;
+import org.runejs.client.sound.Runnable_Impl1;
 import org.runejs.client.GameShell;
-import org.runejs.client.Runnable_Impl1;
 import org.runejs.Configuration;
 
 import java.io.*;
